+++ src/main/java/com/harimahat/customer/CustomerDao.java
@@ -1,7 +1,5 @@
 package com.harimahat.customer;
 
-import jakarta.persistence.criteria.CriteriaBuilder;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -14,17 +12,17 @@ public interface CustomerDao {
 
     List<Customer> selectAllCustomers();
 
-    Optional <Customer> selectCustomerById(Integer customerId);
+    Optional <Customer> selectCustomerById(Long customerId);
 
     void insertCustomer( Customer customer);
 
-    void  deleteCustomerById(Integer customerId);
+    void  deleteCustomerById(Long customerId);
 
     void  updateCustomer(Customer update);
 
     boolean existPersonWithEmail(String emailcustomer);
 
-    boolean existPersonWithId(Integer customerId);
+    boolean existPersonWithId(Long customerId);
